stochatic L-Systems handling + progress on exection environment

+++ linn.core/src/main/java/linn/core/lang/ProductionRuleBuilder.java
@@ -26,7 +26,7 @@ public class ProductionRuleBuilder {
 		return this;
 	}
 	
-	public ProductionRuleBuilder andWeight(float weight) {
+	public ProductionRuleBuilder andWeight(double weight) {
 		this.linn.setRuleWeight(currentRuleId, weight);
 		return this;
 	}
